requirements.txt
google-genai
SpeechRecognition
thefuzz
pyaudio
paramiko
audioop-lts ; python_version >= '3.13'
pyttsx3 ; sys_platform == 'win32'
qi ; sys_platform == 'linux'
pyreadline3
dotenv
flask
hypercorn
pyyaml
git+https://github.com/dsc/bunch.git  # bunch